requirements.txt
Flask==3.0.0
Folium==0.14.0
GeoPandas==0.14.0
requests==2.31.0
geopy==2.4.0
python-dotenv==1.0.0
supabase==2.4.0
